django_api/aia/requirements.txt
autopep8==1.4.4
certifi==2019.6.16
Django==2.2.1
django-filter==2.2.0
django-mongoengine-filter==0.3.4
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.4
djangorestframework-queryfields==1.0.0
docxtpl==0.6.3
Genshi==0.7.3
Jinja2==2.10.1
lxml==4.4.1
Markdown==3.1.1
markdown2==2.3.8
MarkupSafe==1.1.1
mongoengine==0.18.2
pycodestyle==2.5.0
pymongo==3.9.0
python-dateutil==2.8.0
python-docx==0.8.7
pytz==2019.2
relatorio==0.9.0
secretary==0.2.19
six==1.12.0
sqlparse==0.3.0
